+++ airbyte-integrations/connectors/source-twitter-followers/source_twitter_followers/schemas/twitter_tweet_metrics.json
