+++ airbyte-integrations/connectors/source-twitter-followers/source_twitter_followers/spec.yaml
@@ -21,7 +21,8 @@ connectionSpecification:
         - twitterdev
     page_size:
       title: Page Size
-      description: The size of per page
+      description: The size of per pagels
+
       type: integer
       default: 5000
       examples:
