--- airbyte-integrations/connectors/source-twitter-tweet/.dockerignore
@@ -1,6 +1,0 @@
-*
-!Dockerfile
-!main.py
-!source_twitter_tweet
-!setup.py
-!secrets
