--- airbyte-integrations/connectors/source-twitter-tweet/integration_tests/__init__.py
@@ -1,3 +1,0 @@
-#
-# Copyright (c) 2022 Airbyte, Inc., all rights reserved.
-#
