--- airbyte-integrations/connectors/source-twitter-tweet/integration_tests/abnormal_state.json
@@ -1,5 +1,0 @@
-{
-  "todo-stream-name": {
-    "todo-field-name": "todo-abnormal-value"
-  }
-}
