--- airbyte-integrations/connectors/source-twitter-tweet/integration_tests/acceptance.py
@@ -1,16 +1,0 @@
-#
-# Copyright (c) 2022 Airbyte, Inc., all rights reserved.
-#
-
-
-import pytest
-
-pytest_plugins = ("connector_acceptance_test.plugin",)
-
-
-@pytest.fixture(scope="session", autouse=True)
-def connector_setup():
-    """This fixture is a placeholder for external resources that acceptance test might require."""
-    # TODO: setup test dependencies if needed. otherwise remove the TODO comments
-    yield
-    # TODO: clean up test dependencies
