--- airbyte-integrations/connectors/source-twitter-tweet/integration_tests/configured_catalog.json
@@ -1,13 +1,0 @@
-{
-  "streams": [
-    {
-      "stream": {
-        "name": "twitter_tweet_metrics",
-        "json_schema": {},
-        "supported_sync_modes": ["full_refresh"]
-      },
-      "sync_mode": "full_refresh",
-      "destination_sync_mode": "overwrite"
-    }
-  ]
-}
