--- airbyte-integrations/connectors/source-twitter-tweet/integration_tests/invalid_config.json
@@ -1,5 +1,0 @@
-{
-  "api_key": "",
-  "screen_name": "",
-  "page_size": 60
-}
