--- airbyte-integrations/connectors/source-twitter-tweet/integration_tests/sample_config.json
@@ -1,3 +1,0 @@
-{
-  "fix-me": "TODO"
-}
