--- airbyte-integrations/connectors/source-twitter-tweet/integration_tests/sample_state.json
@@ -1,5 +1,0 @@
-{
-  "todo-stream-name": {
-    "todo-field-name": "value"
-  }
-}
