--- airbyte-integrations/connectors/source-twitter-tweet/requirements.txt
@@ -1,2 +1,0 @@
--e ../../bases/connector-acceptance-test
--e .
