--- airbyte-integrations/connectors/source-twitter-tweet/setup.py
@@ -1,29 +1,0 @@
-#
-# Copyright (c) 2022 Airbyte, Inc., all rights reserved.
-#
-
-
-from setuptools import find_packages, setup
-
-MAIN_REQUIREMENTS = [
-    "airbyte-cdk~=0.2",
-]
-
-TEST_REQUIREMENTS = [
-    "pytest~=6.2",
-    "pytest-mock~=3.6.1",
-    "connector-acceptance-test",
-]
-
-setup(
-    name="source_twitter_tweet",
-    description="Source implementation for Twitter Tweet.",
-    author="Airbyte",
-    author_email="contact@airbyte.io",
-    packages=find_packages(),
-    install_requires=MAIN_REQUIREMENTS,
-    package_data={"": ["*.json", "*.yaml", "schemas/*.json", "schemas/shared/*.json"]},
-    extras_require={
-        "tests": TEST_REQUIREMENTS,
-    },
-)
