--- airbyte-integrations/connectors/source-twitter-tweet/source_twitter_tweet/__init__.py
@@ -1,8 +1,0 @@
-#
-# Copyright (c) 2022 Airbyte, Inc., all rights reserved.
-#
-
-
-from .source import SourceTwitterTweet
-
-__all__ = ["SourceTwitterTweet"]
