--- airbyte-integrations/connectors/source-twitter-tweet/source_twitter_tweet/spec.yaml
@@ -1,28 +1,0 @@
-documentationUrl: https://docsurl.com
-connectionSpecification:
-  $schema: http://json-schema.org/draft-07/schema#
-  title: Twitter Spec
-  type: object
-  required:
-    - api_key
-    - screen_name
-  additionalProperties: true
-  properties:
-    api_key:
-      title: API Key Token
-      description: API Key
-      type: string
-      airbyte_secret: true
-    screen_name:
-      title: Screen Name
-      description: Twitter User Screen Name
-      type: string
-      examples:
-        - twitterdev
-    page_size:
-      title: Page Size
-      description: The size of per page
-      type: integer
-      default: 5000
-      examples:
-        - 1000
